+++ app/assets/CHANGELOG.md
@@ -26,6 +26,7 @@
 - fix: text message content size calculation (@ew-sirenko)
 - fix: save DNG files to image gallery (@ShlomoCode)
 - fix(ios, android): prevent transfer error by saving unsupported media formats to folder instead of gallery (@ShlomoCode)
+- fix(android): share to LocalSend sometimes doesn't work
 - fix(android): preserve location metadata when sharing media (@ShlomoCode)
 - fix(macos): prevent Dock icon from briefly appearing during autostart when "Start hidden" is enabled (@ShlomoCode)
 - fix(macos): Dock icon drag-and-drop and Share Extension working again (@ShlomoCode)